+++ scripts/cfg_tool.py
@@ -2,176 +2,190 @@
 #  Author: sascha_lammers@gmx.de
 #
 
-# convert config command to json and back
+import sys
+import argparse
+import libs.config as config
+import time
+
+parser = argparse.ArgumentParser(description="Firmware Configuration Tool")
+parser.add_argument("action", help="action to execute", choices=["display", "modify", "monitor"])
+parser.add_argument('-d', '--data', help="data to read or modify", required=False, default=None)
+parser.add_argument('-V', '--version', help='version of the data provided. For example 2.1.x', type=str, default=None)
+parser.add_argument('-p', '--port', help='serial port', type=str, default=None)
+parser.add_argument('-b', '--baud', help='serial port rate', type=int, default=57600)
+parser.add_argument('-a', '--address', help='dimmer address', type=int, default=None)
+parser.add_argument('-s', '--set', help='set key=value', nargs='+', default=[])
+parser.add_argument('-j', '--json', help='use JSON as output format', action="store_true", default=False)
+
+args = parser.parse_args()
+
 #
-# command:
-# +i2ct=17,89,91
+# Examples
 #
-# response:
-# +REM=v840,i2ct=17a2024b0000904002000000b004ff0d0002cdcc8c3f00000500000020001412
+# python.exe .\scripts\cfg_tool.py modify --version 2.2 --data '+REM=v840,i2ct=17a2024b0000904002000000b004ff0d0002cdcc8c3f00000500000020001412' -s range_begin=1
+# Detected version: 2.2.0
+# Set range_begin to 1
+# +I2CT=17a2024b0000904002000000b004ff0d0002cdcc8c3f00000501000020001412
 #
-# if cfg.get_from_rem_command('+REM=v840,i2ct=17a2024b0000904002000000b004ff0d0002cdcc8c3f00000500000020001412'):
-#     print(cfg.version)
-#     print(cfg.get_json())
-#
-# 2.2.0
+# python.exe .\scripts\cfg_tool.py display --version 2.2 --data '+REM=v840,i2ct=17a2024b0000904002000000b004ff0d0002cdcc8c3f00000500000020001412' -s range_begin=1 -j
+# Detected version: 2.2.0
+# Set range_begin to 1
 # {
 #   "i2c_address": "0x17",
-#   "mem_address": "0xa2",
-#   "options": "00000010",
-#   "max_temp": 75,
-#   ....
-#   "switch_on_count": 18
+# ...
+#   "range_begin": 1,
+# ...
 # }
 #
-#   cfg.set_item('minimum_off_time_ticks', 1000)
-#   print(cfg.get_command())
-#
-# +I2CT=17a2024b0000904002000000b004ff0de803cdcc8c3f00000500000020001412
-#
 
-import re
-import struct
-import json
+def set_items(cfg, items):
+    key = '?'
+    val = '?'
+    try:
+        for pair in items:
+            pair = pair.split('=', 2)
+            key = str(pair[0].strip())
+            val = str(pair[1].strip())
+            cfg.set_item(key, val)
+            print('Set %s to %s' % (key, val))
+    except Exception as e:
+        raise ValueError('failed to set value: %s=%s: %s' % (key, val, str(e)))
 
-class Config:
-    def __init__(self):
-        self.version = 'unknown'
-        self.struct = {
-            'i2c_address': 'B',
-            'mem_address': 'B',
-            'options': 'B',
-            'max_temp': 'B',
-            'fade_in_time': 'f',
-            'temp_check_interval': 'B',
-            '__empty0': 'B',
-            '__empty1': 'B',
-            'halfwave_adjust_ticks': 'b',
-            'zero_crossing_delay_ticks': 'H',
-            'minimum_on_time_ticks': 'H',
-            'minimum_off_time_ticks': 'H',
-            'internal_1_1v_ref': 'f',
-            'int_temp_offset': 'b',
-            'ntc_temp_offset': 'b',
-            'report_metrics_max_interval': 'B',
-            'range_begin': 'h',
-            'range_end': 'h',
-            'switch_on_minimum_ticks': 'H',
-            'switch_on_count': 'B',
-        }
-        self.config = {}
-
-    def get_unpack(self):
-        s = '<'
-        for key, val in self.struct.items():
-            s = s + val
-        return s
-
-    def set_items(self, items):
-        self.config = self.struct.copy();
-        n = 0
-        for key, val in self.config.items():
-            self.config[key] = items[n]
-            n += 1
-
-    def validate_value(self, type, val):
-        if type in 'BbHh' and not isinstance(val, int):
-            raise ValueError('value not integer: %s' % str(val))
-        if type in 'f' and not isinstance(val, float):
-            raise ValueError('value not float: %s' % str(val))
-        if type=='B':
-            if val<0 or val>255:
-                raise ValueError('value out of range: 0 to 255: %d' % val)
-        if type=='b':
-            if val<-128 or val>127:
-                raise ValueError('value out of range: -128 to 127: %d' % val)
-        if type=='H':
-            if val<0 or val>65535:
-                raise ValueError('value out of range: 0 to 65535: %d' % val)
-        if type=='h':
-            if val<-32768 or val>32867:
-                raise ValueError('value out of range: -32868 to 32867: %d' % val)
-        return True
-
-    def set_item(self, key, val):
-        if not key in self.config:
-            raise KeyError('key does not exist: %s' % key)
-        type = self.struct[key]
-        if self.validate_value(type, val):
-            self.config[key] = val
-            return True
-        return False
-
-    def dump(self):
-        names = list(self.struct.keys())
-        for key, val in self.config.items():
-            print('%s = %s' % (key, str(val)))
-
-    def get_json(self):
-        tmp = self.config.copy();
-        tmp['i2c_address'] = '0x%02x' % tmp['i2c_address']
-        tmp['mem_address'] = '0x%02x' % tmp['mem_address']
-        tmp['options'] = '{:08b}'.format(tmp['options'])
-        return json.dumps(tmp, indent=2)
-
-    def get_bytes(self):
-        b = bytearray()
-        for key, val in self.config.items():
-            fmt = self.struct[key]
+if args.action!='monitor':
+    if not args.data and not args.port:
+        parser.error('--data or --port required')
+    if args.data:
+        if not args.version:
+            parser.error('--data requires --version')
+        else:
             try:
-                packed = struct.pack('<' + fmt, val)
+                version = args.version.split('.')
+                if len(version)<2 or len(version)>3:
+                    raise Exception()
+                parts = []
+                for val in version:
+                    parts.append(int(val))
+                if len(parts) == 2:
+                    parts.append('x')
+                version = parts
+            except:
+                parser.error('version format: major.minor[.revision]')
+
+if args.port:
+    try:
+        import serial
+    except:
+        parser.error('cannot import serial. install pyserial: pip install pyserial')
+
+    import libs.serial_protocol as protocol
+
+    monitor = args.action=='monitor'
+
+    ser = serial.Serial(baudrate=args.baud, timeout=5)
+    sp = protocol.Protocol(ser, args.address)
+    ser.port = args.port
+    ser.dtr = False
+    ser.rts = True
+    ser.open()
+    if ser.isOpen():
+        data = sp.read_request(sp.get_version())
+        if data==False:
+            if not monitor:
+                parser.error('serial read timeout')
+        if data==None:
+            print("Reboot detected, waiting 3 seconds...")
+            time.sleep(3)
+            data = sp.read_request(sp.get_version())
+
+        if isinstance(data, bytearray):
+            version = config.Config.get_version(data)
+            print("Detected version: %u.%u.%u" % (version[0], version[1], version[2]))
+            version_key = config.Config.is_version_supported(version[0], version[1], version[2], False)
+            if version_key==False:
+                parser.error('version %s.%s.%s not supported' % (str(version[0]), str(version[1]), str(version[2])))
+            cfg = config.Config(version_key)
+
+        if monitor:
+            print("Monitoring serial port...")
+            while monitor:
+                line = sp.readline()
+                sp.print_data(line)
+        else:
+            sp.get_config()
+            n = 0
+            while n < 10:
+                line = sp.readline()
+                try:
+                    cfg.get_from_rem_command(line)
+                    print('Received config version: %s' % cfg.version)
+                    print('Set command: %s' % cfg.get_command())
+                    n = -1
+                    break
+                except Exception as e:
+                    print('Parse error: %s', str(e))
+                    pass
+                n += 1
+
+            if n!=-1:
+                print('Timeout receiving configuration')
+                sys.exit(-1)
+
+            try:
+                set_items(cfg, args.set)
             except Exception as e:
-                raise TypeError('invalid type for key: %s: type: %s' % (key, type(val)))
-            b = b + packed;
-        return b
+                print(str(e))
+                sys.exit(-1)
 
-    def get_command(self):
-        return '+I2CT=' + self.get_bytes().hex()
+            if args.action=='display':
+                if args.json==True:
+                    print(cfg.get_json())
+                else:
+                    print(cfg.get_command())
+            elif args.action=='modify':
+                data = cfg.get_command().encode()
+                print("Sending configuration: ", data)
+                ser.write(data + b'\n')
+                print("Sending store command...")
+                sp.store_config()
 
-    def get_from_command(self, cmd):
-        m = re.match('^\+i2ct=([0-9a-f]+)', cmd, re.IGNORECASE)
-        print(m)
-        if m:
-            self.version = 'unknown'
-            fmt = self.get_unpack()
-            config = bytearray.fromhex(m[1])
-            self.set_items(struct.unpack(fmt, config[0:struct.calcsize(fmt)]));
-            return True
-        return False
+                n = 0
+                while n < 5:
+                    line = sp.readline()
+                    if sp.print_data(line)==0xf3:
+                        break
+                    n += 1
+                print('Closing connection...')
 
-    def get_from_rem_command(self, cmd):
-        m = re.match('^\+rem=v([0-9a-f]{1,4}),i2ct=([0-9a-f]+)', cmd, re.IGNORECASE)
-        if m:
-            version = m[1]
-            while len(version) < 4:
-                version = '0' + version
-            version = bytearray.fromhex(version)
-            (version, ) = struct.unpack('>H', version)
-            self.version = '%u.%u.%u' % (version >> 10, ((version >> 5) & 0x1f), version & 0x1f)
-            if not self.version.startswith('2.2.'):
-                return False
+            ser.close()
 
-            fmt = self.get_unpack()
-            config = bytearray.fromhex(m[2])
-            self.set_items(struct.unpack(fmt, config[0:struct.calcsize(fmt)]));
-            return True
+else:
+    version_key = config.Config.is_version_supported(version[0], version[1], version[2], False)
+    if version_key==False:
+        parser.error('version %s.%s.%s not supported' % (str(version[0]), str(version[1]), str(version[2])))
+    cfg = config.Config(version_key)
 
-        return False
+    try:
+        cfg.get_from_rem_command(args.data)
+        print('Detected version: %s' % cfg.version)
+    except Exception as e:
+        try:
+            cfg.get_from_command(args.data)
+        except Exception as e:
+            parser.error('invalid data: %s' % str(e))
 
+    if len(args.set):
+        if args.json==None:
+            args.json = False
+        try:
+            set_items(cfg, args.set)
+        except Exception as e:
+            parser.error(str(e))
 
-cfg = Config()
+    if args.address!=None:
+        cfg.set_item('i2c_address', args.address)
 
-line = '+REM=v840,i2ct=17a2024b0000904002000000b004ff0d0002cdcc8c3f00000500000020001412\n'
-if cfg.get_from_rem_command(line):
-    # print(cfg.version)
-    # print(cfg.get_json())
-
-    cfg.set_item('minimum_off_time_ticks', 1000)
-
-    print(cfg.get_command())
-
-# line = '+i2ct=17a2024b0000904002000000b004ff0d0002cdcc8c3f00000500000020001412\n'
-# if cfg.get_from_command(line):
-#     print(cfg.get_json())
-#     print(cfg.version)
+    if args.json==True:
+        print(cfg.get_json())
+    else:
+        print(cfg.get_command())